Backend Services skeleton created! within backend_services folder we have - userService -> containing account and user data, methods, endpoints - postService -> containing post(content) data - creatorService -> containing creator data - subscriptionService -> containing subscription and plan data This structure should follow Sam's database design. User service is the only fleshed out service atm - has CRUD operations for user, however account logic yet to be done!
