+++ backend_services/creatorService/src/main/resources/application.properties
@@ -1,0 +1,15 @@
+# Server configuration different port for each service
+server.port=8083
+
+# Database configuration
+spring.datasource.url=jdbc:mysql://localhost:(port local db is on)/(db name)
+spring.datasource.username= db username
+spring.datasource.password= db pass
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+# this driver is needed to translate java database api calls to sql queries that the db can understand
+
+
+# JPA configuration
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
