+++ backend_services/postService/src/main/java/onlypans/postService/PostServiceApplication.java
@@ -1,0 +1,11 @@
+package onlypans.postService;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PostServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(PostServiceApplication.class, args);
+    }
+}
