+++ backend_services/postService/src/main/java/onlypans/postService/controller/PostController.java
@@ -1,0 +1,35 @@
+package onlypans.postService.controller;
+
+import onlypans.postService.entity.Post;
+import onlypans.postService.service.PostService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/posts")
+public class PostController {
+    // api endpoints for post application
+    @Autowired
+    private PostService postService;
+
+    @PostMapping
+    public ResponseEntity<Post> createPost(@RequestBody Post post) {
+        return new ResponseEntity<>(postService.createPost(post), HttpStatus.CREATED);
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<Post> getPostById(@PathVariable Long id) {
+        return postService.getPostById(id)
+                .map(post -> new ResponseEntity<>(post, HttpStatus.OK))
+                .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
+    }
+
+    @GetMapping
+    public List<Post> getAllPosts() {
+        return postService.getAllPosts();
+    }
+}
