+++ backend_services/subscriptionService/src/main/java/onlypans/subscriptionService/SubscriptionServiceApplication.java
@@ -1,0 +1,4 @@
+package onlypans.subscriptionService;
+
+public class SubscriptionServiceApplication {
+}
