+++ backend_services/subscriptionService/src/main/java/onlypans/subscriptionService/controller/PlanController.java
@@ -1,0 +1,4 @@
+package onlypans.subscriptionService.controller;
+
+public class PlanController {
+}
