+++ backend_services/subscriptionService/src/main/java/onlypans/subscriptionService/controller/SubscriptionController.java
@@ -1,0 +1,4 @@
+package onlypans.subscriptionService.controller;
+
+public class SubscriptionController {
+}
