+++ backend_services/subscriptionService/src/main/java/onlypans/subscriptionService/entity/Plan.java
@@ -1,0 +1,4 @@
+package onlypans.subscriptionService.entity;
+
+public class Plan {
+}
