+++ backend_services/subscriptionService/src/main/java/onlypans/subscriptionService/repository/PlanRepository.java
@@ -1,0 +1,4 @@
+package onlypans.subscriptionService.repository;
+
+public interface PlanRepository {
+}
