+++ backend_services/subscriptionService/src/main/java/onlypans/subscriptionService/repository/SubscriptionRepository.java
@@ -1,0 +1,4 @@
+package onlypans.subscriptionService.repository;
+
+public interface SubscriptionRepository {
+}
