+++ backend_services/subscriptionService/src/main/java/onlypans/subscriptionService/service/PlanService.java
@@ -1,0 +1,4 @@
+package onlypans.subscriptionService.service;
+
+public class PlanService {
+}
