+++ backend_services/userService/src/main/java/onlypans/userService/UserServiceApplication.java
@@ -1,0 +1,11 @@
+package onlypans.userService;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(UserServiceApplication.class, args);
+    }
+}
