+++ backend_services/userService/src/main/java/onlypans/userService/entity/User.java
@@ -1,0 +1,34 @@
+package onlypans.userService.entity;
+
+import jakarta.persistence.*;
+
+
+@Entity
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String username;
+    private String email;
+    private String password;
+    private String stripeId;
+
+    @OneToOne(mappedBy = "user")
+    private Account account;
+
+
+    // Getters and setters
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
+
+    public String getUsername() { return username; }
+    public void setUsername(String username) { this.username = username; }
+
+    public String getEmail() { return email; }
+    public void setEmail(String email) { this.email = email; }
+
+    public String getPassword() { return password; }
+    public void setPassword(String password) { this.password = password; }
+}
