+++ backend_services/userService/src/main/java/onlypans/userService/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package onlypans.userService.repository;
+
+import onlypans.userService.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+  // custom queries?
+}
