--- postService/src/main/java/com/example/onlypans/postService/PostServiceApplication.java
@@ -1,11 +1,0 @@
-package com.example.onlypans.postService;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PostServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(PostServiceApplication.class, args);
-    }
-}
