--- postService/src/main/java/com/example/onlypans/postService/service/PostService.java
@@ -1,34 +1,0 @@
-package com.example.onlypans.postService.service;
-
-import com.example.onlypans.postService.entity.Post;
-import com.example.onlypans.postService.repository.PostRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-
-@Service
-public class PostService {
-// all methods the controller will use
-    @Autowired
-    private PostRepository postRepository;
-
-    public Post createPost(Post post) {
-        return postRepository.save(post);
-    }
-
-    public Optional<Post> getPostById(Long id) {
-        return postRepository.findById(id);
-    }
-
-    public List<Post> getAllPosts() {
-        return postRepository.findAll();
-    }
-
-
-
-    public void deletePost(Long id) {
-        postRepository.deleteById(id);
-    }
-}
