--- userService/src/main/java/com/example/onlypans/userService/UserServiceApplication.java
@@ -1,11 +1,0 @@
-package com.example.onlypans.userService;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(UserServiceApplication.class, args);
-    }
-}
