--- userService/src/main/java/com/example/onlypans/userService/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.onlypans.userService.repository;
-
-import com.example.onlypans.userService.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-  // custom queries?
-}
